# [Enhancement] Optimize ingestion performance for table with materialized index STEP 1