… be/src/exec/tablet_sink.cpp
… be/src/runtime/load_channel_mgr.cpp
… be/src/runtime/local_tablets_channel.cpp
… be/src/runtime/tablets_channel.h
… be/src/service/internal_service.cpp
… be/src/service/internal_service.h
… be/src/service/service_be/internal_service.cpp
… be/src/service/service_be/internal_service.h
… be/test/runtime/lake_tablets_channel_test.cpp
… gensrc/proto/internal_service.proto