## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #7717

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->
When a table has multiple materialized views, StarRocks will send the chunks to the tablets corresponding to the materialized views, which leads to the enlargement of network transmission bandwidth and unnecessary data serialization and transmission overhead. Since the materialized view and the base table use the same partition column and distribute in the same node when using `colocate mv` #7451, we can reuse the chunk data of the base table.


| materialized index num| main branch | optimize branch|SpeedUp|
|---|---|---|---|
|0|1314MB/s|1310MB/s||
|1|780MB/s|1274Mb/s|1.6x|
|3|420MB/s|1241MB/s|2.9x|
| 5 |286MB/s |1213MB/s|4.2x|

> using [github_events](https://ghe.clickhouse.tech/) data, 16 parallel
> 3 BE with 1.5GB/s network bandwidth
> create two columns of materialized view; materialized view computation is relatively small; ensure the bottleneck is network bandwidth and serialization.

`NOTE` We will enable it after colocate mv index implementation.