@meegoo 2022-07-25
run starrocks_be_unittest