@meegoo 2022-07-28
run starrocks_fe_unittest