@meegoo 2022-08-12
run starrocks_fe_unittest