@wanpengfei-git 2022-08-16
run starrocks_admit_test